requirements.txt
asgiref==3.9.2
Django==5.2.6
django-debug-toolbar==6.0.0
djangorestframework==3.16.1
drf-yasg==1.21.11
Faker==37.12.0
inflection==0.5.1
iniconfig==2.3.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
sqlparse==0.5.3
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
